Add "limit" and "offset" support for queries Add tests for meta.totalCount

+++ revcrm-server/src/main/kotlin/org/revcrm/graphql/EntitySearchResults.kt
@@ -6,5 +6,5 @@ class EntitySearchResults<T>(
 )
 
 class EntitySearchMeta(
-    val totalCount: Int
+    val totalCount: Long
 )
